// comunicacionesDirectas/src/shared/vendor/tracasa-real/components/arbol-relacionados/arbol-relacionados.module.scss.d.ts
export type Styles = {
  arbol_relacionados__item: string;
  arbol_relacionados__list: string;
  arbol_relacionados__organo_judicial: string;
  arbol_relacionados__seleccionar_todos: string;
  arbol_relacionados__toggle_button: string;
  expediente__activo: string;
  expediente__body: string;
  expediente__checkbox: string;
  expediente__contenedor: string;
  expediente__fila: string;
  expediente__link: string;
  expediente__separador: string;
  expediente__sin_acceso: string;
  expediente__toggle_button: string;
  nivel_0: string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;